mv RethEvmBuilder to a builder module
